refactor(jsx): move expression-parser to @barefootjs/jsx and use ParsedExpr AST in filterPredicate
- Move expression-parser.ts from go-template to jsx package
- Update IRLoop.filterPredicate to use ParsedExpr instead of string
- Early parse filter predicates in jsx-to-ir.ts with isSupported() check
- Remove redundant re-parse in go-template-adapter
- Export parseExpression, isSupported, exprToString from @barefootjs/jsx
